kobrabitescore/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.4
django-bootstrap5==24.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-db-mutex==3.1.0
django-filter==24.3
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
filelock==3.16.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
prompt_toolkit==3.0.48
psycopg2==2.9.10
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
tldextract==5.1.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
zope.event==5.0
zope.interface==7.2
